+++ src/engine.cxx
@@ -2,15 +2,15 @@
 #include <iostream>
 
 namespace FEngine {
+
 	Engine::Engine() {
 		// Construct logger
 		this->logger_ = FLogger::GetInstance();
-
+		
+		FLogger::LoadConfiguration("loggers.cfg");
+		LOG(INFO) << "========== LOG START  ==========";
 	}
 	Engine::~Engine() {
 
 	}
-
-	//template<>
-	//std::shared_ptr<Engine> Utils::Singleton<Engine>::instance = nullptr;
 }