+++ src/utils/f_typedefs.hpp
@@ -1,0 +1,30 @@
+#ifndef _3FENGINE_SRC_UTILS_TYPEDEFS_HPP_
+#define _3FENGINE_SRC_UTILS_TYPEDEFS_HPP_
+
+#include <memory>
+#include <string>
+
+#include "f_vertices.hpp"
+
+namespace FEngine {
+#define F_DISABLE_COPY(Class) \
+	Class(const Class&) = delete; \
+	Class& operator=(const Class&) = delete;
+
+	template<class T>
+	using FShared = std::shared_ptr<T>;
+
+	template<class T>
+	using FUnique = std::unique_ptr<T>;
+
+	using FString = std::string;
+
+	using FVertices3f = FVertices<float, 3>;
+	using FVertex3f = FVertices3f::Vertex;
+
+	using FIndices3 = FVertices<int, 3>;
+	using FUVsf = FVertices<float, 2>;
+
+}
+
+#endif