+++ src/utils/f_utils.cxx
@@ -1,9 +1,9 @@
-#include "utils.hpp"
+#include "f_utils.hpp"
 #include <algorithm>
 
 namespace FEngine {
-	namespace Utils {
-		std::string ToLower(const std::string & str)
+	namespace FUtils {
+		FString ToLower(const FString & str)
 		{
 			auto lower_case = str;
 			std::transform(lower_case.begin(), lower_case.end(), lower_case.begin(), ::tolower);