+++ src/utils/f_utils.hpp
@@ -1,0 +1,11 @@
+#ifndef _3FENGINE_SRC_UTILS_F_UTILS_HPP_
+#define _3FENGINE_SRC_UTILS_F_UTILS_HPP_
+
+#include "f_typedefs.hpp"
+namespace FEngine {
+	namespace FUtils {
+		std::string ToLower(const FString & str);
+	}
+}
+
+#endif  // _3FENGINE_SRC_UTILS_F_UTILS_HPP_