+++ src/utils/f_vertices.cxx
@@ -1,0 +1,5 @@
+#include "f_vertices.hpp"
+
+namespace FEngine {
+
+}