+++ src/utils/utils.hpp
@@ -1,7 +1,0 @@
-#pragma once
-#include <string>
-namespace FEngine {
-	namespace Utils {
-		std::string ToLower(const std::string & str);
-	}
-}