+++ src/utils/vertices.cxx
@@ -1,5 +1,0 @@
-#include "vertices.hpp"
-
-namespace FEngine {
-
-}